requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
streamlit>=1.29.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
python-docx>=1.0.0
requests>=2.31.0
tiktoken>=0.5.0
google-generativeai==0.8.5
-e .